# openapi.json
{
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "properties": {
          "atPeriodEnd": {
            "default": true,
            "title": "Atperiodend",
            "type": "boolean"
          }
        },
        "title": "CancelRequest",
        "type": "object"
      },
      "CancelResponse": {
        "description": "Written to be shown *before* the user confirms \u2014 contract \u00a77.",
        "properties": {
          "accessUntil": {
            "format": "date-time",
            "title": "Accessuntil",
            "type": "string"
          },
          "cancelAtPeriodEnd": {
            "title": "Cancelatperiodend",
            "type": "boolean"
          },
          "creditsKept": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Creditskept",
            "type": "object"
          },
          "creditsLostAtPeriodEnd": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Creditslostatperiodend",
            "type": "object"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "plan",
          "status",
          "cancelAtPeriodEnd",
          "accessUntil",
          "creditsLostAtPeriodEnd",
          "creditsKept"
        ],
        "title": "CancelResponse",
        "type": "object"
      },
...
      "CheckoutRequest": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "maxLength": 3,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "returnUrl": {
            "maxLength": 2048,
            "title": "Returnurl",
            "type": "string"
          }
        },
        "required": [
          "plan",
          "returnUrl"
        ],
        "title": "CheckoutRequest",
        "type": "object"
      },
      "CheckoutResponse": {
        "description": "What a plan change produced.\n\n**`checkoutUrl` is nullable, which is an amendment to contract \u00a77.** The\nendpoint is specified as *\"start a subscription or change plan\"*, and one\nkind of plan change costs nothing now: a downgrade applies at the next\nperiod boundary (\u00a78.3) so nobody loses credits they are half way through\nusing. There is no page to pay on, and sending the user to one would be\nasking them to pay in order to receive less.\n\nReporting that as an error was the first shape this took, and it was wrong\ntwice: a scheduled downgrade is a success, and \u2014 worse \u2014 the state change\nwas written and then thrown away, because `get_session` rolls back on any\nexception. The client now has one path for \"your plan changed\": follow\n`checkoutUrl` if there is one, otherwise show `effectiveAt`.",
        "properties": {
          "checkoutUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkouturl"
          },
          "effectiveAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectiveat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "scheduledPlan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledplan"
          }
        },
        "title": "CheckoutResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortalRequest": {
        "properties": {
          "returnUrl": {
            "maxLength": 2048,
            "title": "Returnurl",
            "type": "string"
          }
        },
        "required": [
          "returnUrl"
        ],
        "title": "PortalRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI Video Editor. The contract for this API is documented in docs/05-api-contract.md; this schema is generated from it and is the source the frontend generates its types from.",
    "title": "ZipZop API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/cancel": {
      "post": {
        "description": "One request, no funnel, and the cost stated in the response.\n\nThe body is written to be shown to the user *before* they confirm: losing\n1,840 credits is worth knowing about in advance, and saying so plainly at\nthe moment of cancelling is both fairer and better retention than\ndiscovering it a week later. There is deliberately no retention offer here\nand no extra step \u2014 the competitor research that prompted this milestone was\nabout exactly that (docs/20-m6-readiness.md \u00a75).",
        "operationId": "cancel_v1_billing_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop the subscription renewing",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/checkout": {
      "post": {
        "description": "A hosted page to pay on. **The provider is derived, never chosen.**\n\nA downgrade does not go through checkout at all: it is scheduled for the\nnext period boundary so nobody loses credits they are half way through\nusing (\u00a78.3), and sending the user to a payment page for it would be asking\nthem to pay to receive less.",
        "operationId": "checkout_v1_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a subscription or change plan",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/portal": {
      "post": {
        "description": "`portalUrl` may be null, and that is a real answer.\n\nRazorpay hosts no customer portal. The client shows what *is* available \u2014\ncancel here, invoices by email \u2014 rather than opening a page that does not\nexist. Amendment recorded in contract \u00a77.",
        "operationId": "portal_v1_billing_portal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manage the subscription on the provider's own page",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/topup": {
      "post": {
        "operationId": "topup_v1_billing_topup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Buy credits that never expire",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/topup-packs": {
      "get": {
        "description": "What `POST /billing/topup` will accept.\n\nContract \u00a77 names `credits_5000` in its example without saying where the\nlist comes from. A client that has to hardcode pack codes is a client that\nbreaks the day a pack is renamed.",
        "operationId": "list_topup_packs_v1_billing_topup_packs_get",
        "parameters": [
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopupPacksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Credit packs that never expire",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/credits/ledger": {
      "get": {
        "description": "A first-class endpoint, not an admin tool.\n\nThis is what a support conversation about *\"where did my credits go\"* is\nanswered from, and it is the reason the ledger records a bucket on every\nrow rather than a single balance.",
        "operationId": "credit_ledger_v1_credits_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Every credit movement, with its bucket",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/plans": {
      "get": {
        "description": "Public \u2014 no authentication.\n\nDeliberately so: a pricing page that needs an account to show a price asks\npeople to sign up before they know what it costs.\n\n`suggestedCurrency` comes from whatever the edge network worked out about\nthe caller's country. **It is a suggestion, not a decision** \u2014 the client\nmust let the user change it with `?currency=`, because VPNs, travellers and\nexpatriates make IP unreliable.",
        "operationId": "list_plans_v1_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlansResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The plans anyone can buy",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/promo/{code}": {
      "get": {
        "description": "Public, and checked at the sign-up form rather than after registering.\n\nThe attribution is written once and never revisited, so a mistyped code\ndiscovered afterwards can never be applied. A message under the field costs\none request; the alternative costs a customer and a commission.\n\nAlways `200`. A 404 here would let anyone enumerate which codes exist, and\nthe client's rendering is the same either way.",
        "operationId": "preview_promo_v1_promo__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Is this code real, and what does it give?",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/promo/{code}/stats": {
      "get": {
        "description": "Only the owner sees it.\n\nA code's numbers say how many people from a particular Discord server signed\nup and how many pay us \u2014 which is commercial information about somebody\nelse's community, not ours to publish. 404 rather than 403 for a code\nsomebody else owns, so the response cannot be used to discover that a code\nexists.",
        "operationId": "promo_stats_v1_promo__code__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What this code has brought in, for whoever owns it",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}